Final React Changes

+++ src/components/Country.jsx
@@ -6,7 +6,7 @@ import Medal from './Medal';
 import { TrashFill, Save, ArrowCounterclockwise } from 'react-bootstrap-icons';
 
 const Country = (props) => {
-  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset } = props;
+  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset, canDelete, canPatch } = props;
   const getMedalsTotal = () => {
     let sum = 0;
     // use medal count displayed in the web page for medal count totals
@@ -45,7 +45,7 @@ const Country = (props) => {
                 className="icon-btn" />
             </React.Fragment>
             :
-            <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
+            canDelete && <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
           }
         </Card.Title>
         <ListGroup variant="flush">
@@ -54,6 +54,7 @@ const Country = (props) => {
             <Medal  
               country={ country } 
               medal={ medal } 
+              canPatch={ canPatch }
               onIncrement={ onIncrement } 
               onDecrement={ onDecrement } />
           </ListGroup.Item>
